split code in index.js into smaller files
