+++ js/api/constants.js
@@ -1,0 +1,9 @@
+// All API keys. Export so the can be imported in other modules
+export const API_KEY = "0ae6fc73-5bc5-4996-b33f-99391a812764";
+
+export const API_BASE = "https://v2.api.noroff.dev";
+export const API_AUTH = "/auth";
+export const API_REGISTER = "/register";
+export const API_LOGIN = "/login";
+export const API_KEY_URL = "/create-api-key";
+export const API_POSTS = "/social/posts";
