+++ js/storage/load.js
@@ -1,0 +1,4 @@
+export function load(key) {
+	return JSON.parse(localStorage.getItem(key));
+}
+
