+++ js/storage/save.js
@@ -1,0 +1,4 @@
+//local storage functions
+export function save(key, value) {
+	localStorage.setItem(key, JSON.stringify(value));
+}
